refactor: move sprint-game folder into pages
