+++ src/core/components/app/app.ts
@@ -1,4 +1,4 @@
-import { SprintGameApp } from '../sprint-game/app/sprint-game-app';
+import { SprintGameApp } from '../../../pages/sprint-game/app/sprint-game-app';
 import AudioCallApp from '../audio-call/app/audio-call-app';
 export default class App {
   public sprintGameApp: SprintGameApp;
