+++ src/index.ts
@@ -2,7 +2,7 @@ import './global.css';
 
 import WordListPage from './pages/wordlist/wordlist-page';
 import AuthPage from './pages/auth/auth-page';
-import { SprintGameApp } from './core/components/sprint-game/app/sprint-game-app';
+import { SprintGameApp } from './pages/sprint-game/app/sprint-game-app';
 import AudioCallApp from './core/components/audio-call/app/audio-call-app';
 // import App from './core/components/app/app';
