+++ src/pages/sprint-game/sprint-game-view/sprint-game-view.ts
