+++ src/pages/sprint-game/timer/timer.css
